Add disable_maintenance parameter to remove cronjob if unwanted
Prior to this commit you would get the cron job installed but to
disable the maintenance cron job you'd have to remove the class
and remove the cron manually.

With the addition of this commit, you can simply set the disable
maintenance parameter to remove the cron job for you.