Break out analyze_scale_factor and vacuum_scale_factor params
Prior to this commit, these two settings were based on the same
single parameter.

After this commit, the setting have their own dedicated parameter.

In a default install of postgresql the analyze scale factor is .1
and the vacuum scale factor is .2.  I assume the ratio is correct
so when reducing the settings to be more aggressive I keep the
2 to 1 ratio of the settings.