Add prototype defined type for backing up databases
This commit adds the ability to make one cron job per database
to run pg_dump.

I don't particularly like the approach as you have to space out the
cron jobs far enough to make sure they won't overlap but then you
also have lots of wasted time in the middle.

In a future commit, it would be better to have a single script to
perform all the backups and have it run via a single cron job.