Add commonly changed postgresql settings as parameters
Prior to this commit, we only managed autovacuum settings.

After this commit, we manage many more commonly changed postgresql
settings.

We also set autovacuum_cost_delay=0 on the fact_values table of
pe-puppetdb because it prevents bloat of the table and its indexes
if customers have a lot of facts that change every run.