Make a defined type for setting table attributes
Prior to this commit, there was a psql command to change a table
attribute.  This was difficult when trying to point people to an
example that would allow them to do the same for a different table

After this commit, the psql command has been abstracted into a
defined type that makes it easier to set arbitrary table attributes.
A new defined type has been made for the specific case of setting
autovacuum_cost_delay=0