Add backup script and defined type for using it
Prior to this commit, there was no backup script.

After this commit, there is a backup script that will iterate
through an array of database names to produce a script to backup
each database.

Using the defined type, you can backup an arbitrary set of
databases on a different cron schedule.