Change ensure_resource(file) back to file resources
Prior to this commit, we used ensure_resource because it was meant
to be inside a defined type that may be specified more than once.
Now the defined type is gone and backup is a class

After this commit, we use file resources for clarity since the class
should not be defined more than once.