Add retention policy to backup script
Prior to this commit, the backup script simply backed up but never
worried about preventing the storage of too many backups.

After this commit, the backup script has a retention policy and
deletes more than the desired number of backups minus 1 backups
so that the new backup can be completed.

This will likely cause valid backups to be deleted if the backups
fail and create an empty file.  Then there will be no valid backups
if nothing is done to intervene but we're choosing that option
over running out of disk space by not aggresively deleting backups.