Reduce factsets autovacuum vacuum scale factor
Prior to this commit, factsets would have the same autovacuum
vacuum scale factor as the global setting.  With this setting we
see that factsets is vacuumed more frequently than any other table
and rarely actually cleans up dead tuples.

After this commit, we reduce the factsets autovacuum vacumm scale
factor to 0.80 (10x the default vacuum scale factor) in an effort to
reduce such frequent vacuuming of the table.  Note that this does
not affect how often the table is analyzed so the only potential
negative impact should be increased growth of the table to store
a few more dead tuples between vacuums.