Make reports autovacuum happen more often and more agressively
Prior to this commit, the reports table used the global autovacuum
settings.

After this commit, we move to a 1% autovacuum_vacuum_scale_factor
so that it is autovacuumed more often.  We also move to a cost_delay
of zero so that autovacuum won't defer to other queries.  This should
prevent the toast table associated with reports from continually
growing.