Make cronjobs to vacuum full puppetdb tables
We want to VACUUM FULL all the tables except reports and
resource events.  We do this across days and for some lesser used
tables we do this only once per month.

By default the script will pause for 5 minutes between VACUUM
FULL of each table.  This allows the PuppetDB queue to work
down from the last VACUUM FULL.

Note that cronjob default times are chosen to note conflict with
the database backup times in the backup manifest.