Remove verbose output and stop redirecting output
Prior to this commit, we included --verbose on the vacuumdb
command so that we could see the command was working.  It turns
out that verbose output goes to STDERR and thus can't be
redirected to a file seperate from actual error messages.

After this commit, we no longer try to redirect output to any files.
Instead we remove the verbose output and hope that when an error
occurs it'll actually fire off a cron email.