Add analyze when performing VACUUM FULL
Prior to this commit, we only did a VACUUM FULL.  However, this
already requires reading the whole table so it's advantageous to
also analyze the table at the same time.  Otherwise autovacuum
will come around and analyze the table later when it could have
already been complted during the VACUUM FULL.

After this commit, we add --analyze to vacuumdb.