(maint) Add the default facts to the .sync.yml
Prior to this commit, there were facts required for spec tests that were
manually added to the default_facts.yaml. These would be overwritten on
a pdk update as they were not in the .sync.yml. This commit adds those
facts to the .sync.yml so they persist pdk updates.