(FEAT) Allow parameter for vacuum full
Prior to this commit, if pg_repack was available
we always used it as it was assumed to be better.

After this commit, we allow users to specify if they'd
like to use vacuum full instead for cases where pg_repack
is failing.  Since we no longer need to repack the
resource_events or reports tables running a vacuum
full on all of the smaller tables is likely not a
large problem.