Fix managing backup class and resources
Prior to this commit, the ensure for the cron entries in the backup
class were hard coded to 'present' and could not be removed once set.
This commit adds a new parameter to manage the cron jobs based on if
$manage_database_backups is set.

Restructure if/else logic

Prior to this commit, the checks for managing the databases and having
systemd were done in the same check, so the warning was confusing.  This
commit moves $manage_database_maintenance to its own if clause.  It also
moves the $manage_database_backups inside the version check if, because
it makes more sense to put all of the resources behind the version
check.