(SUP-2557) Ensure backup class is not included
Prior to this commit, by default the pe_databases::backup class would be
included but not do anything.  That was because the logic for the class
parameter was correct

disable_maintenance => ! $manage_database_backups

But the logic to include the class was not

if defined('$manage_database_backups')

This commit fixes this logic and adds spec tests to cover it.  I also
moved the test for checking the cron resources from an acceptance test
to spec.